+++ src/Systemic/Context.php
@@ -9,11 +9,9 @@ declare(strict_types=1);
 
 namespace DecodeLabs\Systemic;
 
-use DecodeLabs\Systemic\Plugins\Locale as LocalePlugin;
 use DecodeLabs\Systemic\Plugins\Os;
 use DecodeLabs\Systemic\Plugins\Os\Base as OsBase;
 use DecodeLabs\Systemic\Plugins\Process as ProcessPlugin;
-use DecodeLabs\Systemic\Plugins\Timezone as TimezonePlugin;
 
 use DecodeLabs\Veneer\LazyLoad;
 use DecodeLabs\Veneer\Plugin;
@@ -21,10 +19,6 @@ use DecodeLabs\Veneer\Plugin\Wrapper;
 
 class Context
 {
-    #[Plugin]
-    #[LazyLoad]
-    public LocalePlugin $locale;
-
     /**
      * @phpstan-var Os|Wrapper<Os>
      */
@@ -35,8 +29,4 @@ class Context
     #[Plugin]
     #[LazyLoad]
     public ProcessPlugin $process;
-
-    #[Plugin]
-    #[LazyLoad]
-    public TimezonePlugin $timezone;
 }