+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Moved timezone and locale plugins to Cosmos
+
 ## v0.10.5 (2022-11-27)
 * Reverted PTY usage